refactor: separate codes to expression_or_block.rs
