+++ crates/psl/src/codegen/impls/mod.rs
@@ -1,5 +1,6 @@
 mod declarations;
 mod expressions;
+mod fragments;
 mod item;
 mod program;
 mod statements;
